+++ src/extension.ts
@@ -1,5 +1,6 @@
 import * as vscode from "vscode";
 import {
+  applySchemeFilterCommand,
   buildCommand,
   cleanCommand,
   debuggingBuildCommand,
@@ -9,6 +10,7 @@ import {
   generateBuildServerConfigCommand,
   launchCommand,
   openXcodeCommand,
+  pauseSchemeFilterCommand,
   refreshSchemesCommand,
   removeBundleDirCommand,
   resolveDependenciesCommand,
@@ -138,6 +140,7 @@ export function activate(context: vscode.ExtensionContext) {
   const destinationsTreeProvider = new DestinationsTreeProvider({
     manager: destinationsManager,
   });
+  _context.buildTreeProvider = buildTreeProvider;
 
   // Shortcut to push disposable to context.subscriptions
   const d = _context.disposable.bind(_context);
@@ -171,6 +174,8 @@ export function activate(context: vscode.ExtensionContext) {
   d(command("sweetpad.build.diagnoseSetup", diagnoseBuildSetupCommand));
   d(command("sweetpad.build.stop", stopSchemeCommand));
   d(command("sweetpad.build.switchWorktree", switchWorktreeCommand));
+  d(command("sweetpad.build.pauseSchemeFilter", pauseSchemeFilterCommand));
+  d(command("sweetpad.build.applySchemeFilter", applySchemeFilterCommand));
 
   // Testing
   d(command("sweetpad.testing.buildForTesting", buildForTestingCommand));